fix: lngLat error in neighborhood finder
After updating flatbush.json, encountered an error message
related to the helper function that was calculating bounding box.

This commit removes that helper and replaces it with a Turf's
bounding box function -- more resilient and simplifies the code
here.

Note:  also found why the map on neighborhood finder isn't working -
the MAPBOX_ACCESS_TOKEN is referenced using a different name in the
docs, so assuming the Heroku environment variable is misnamed.