Fix map bounds geometry error
after switching to use @turf/bbox, something in the geometry
calculation on bounds was erroring out.  Switched so that
the map will always just use flatbush.json as bounding box.

Left a fixme: because it would be nice to recenter the map on the
address searched, but not urgent right now.